remove old footer and updated footer path
